Refactor backend structure and implement caching
- Moved logger and cache modules to src/lib for better organization.
- Updated test files to reflect new paths for logger, cache, and routes.
- Created schemas for PtcgSet and TcgpSet models in src/models.
- Implemented ping, ptcg-sets, and tcgp-sets routes in src/routes.
- Added MongoDB connection and seeding logic in src/services/db.
- Introduced caching mechanism for ptcgSets and tcgpSets.
- Updated server initialization to use new route structure and logger.
