+++ backend/src/routes/ping.js
