+++ backend/test/sets.test.js
@@ -3,18 +3,18 @@ const { makeAppWithRoute } = require('./utils/helpers');
 
 describe('PTCG and TCGP sets routes (consolidated)', () => {
   afterEach(() => {
-    const cache = require('../cache');
+    const cache = require('../src/lib/cache');
     cache.ptcgSets = [];
     cache.tcgpSets = [];
   });
 
   describe('ptcg-sets', () => {
     it('returns all sets and filters/selects/orders', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [
         { name: 'Alpha Set', code: 'A1', releaseDate: '2020-01-01' },
         { name: 'Beta Set', code: 'B1', releaseDate: '2021-05-10' },
-      ], '../routes/ptcg-sets');
+      ], '../src/routes/ptcg-sets');
 
       const res = await request(app).get('/api/tcgdex/ptcg-sets/');
       expect(res.status).toBe(200);
@@ -31,10 +31,10 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     });
 
     it('returns 500 when cache missing and 400 for invalid regex', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [
         { name: 'Alpha Set', code: 'A1' }
-      ], '../routes/ptcg-sets');
+      ], '../src/routes/ptcg-sets');
 
       cache.ptcgSets = undefined;
       const res = await request(app).get('/api/tcgdex/ptcg-sets/');
@@ -46,11 +46,11 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     });
 
     it('select includes missing fields (undefined) and orderBy works for ptcg', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [
         { name: 'Alpha Set', code: 'A1', releaseDate: '2020-01-01', rank: 2 },
         { name: 'Beta Set', code: 'B1', releaseDate: '2021-05-10', rank: 1 },
-      ], '../routes/ptcg-sets');
+      ], '../src/routes/ptcg-sets');
 
       const res = await request(app).get('/api/tcgdex/ptcg-sets').query({ select: 'name,nonexistent' });
       expect(res.status).toBe(200);
@@ -65,14 +65,14 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     });
 
     it('sort comparator exercises both branches for ptcg orderBy', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
 
       // Make array where first comparison yields true and another yields false during sort
       const app1 = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [
         { name: 'A', rank: 3 },
         { name: 'B', rank: 2 },
         { name: 'C', rank: 1 },
-      ], '../routes/ptcg-sets');
+      ], '../src/routes/ptcg-sets');
 
       const res1 = await request(app1).get('/api/tcgdex/ptcg-sets').query({ orderBy: 'rank' });
       expect(res1.status).toBe(200);
@@ -82,7 +82,7 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
         { name: 'C', rank: 1 },
         { name: 'B', rank: 2 },
         { name: 'A', rank: 3 },
-      ], '../routes/ptcg-sets');
+      ], '../src/routes/ptcg-sets');
 
       const res2 = await request(app2).get('/api/tcgdex/ptcg-sets').query({ orderBy: 'rank' });
       expect(res2.status).toBe(200);
@@ -95,11 +95,11 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
 
   describe('tcgp-sets', () => {
     it('returns tcgp sets, filters and orders', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/tcgp-sets', cache, 'tcgpSets', [
         { name: 'Delta Set', code: 'D1', releaseDate: '2018-03-03' },
         { name: 'Epsilon Set', code: 'E1', releaseDate: '2022-02-02' },
-      ], '../routes/tcgp-sets');
+      ], '../src/routes/tcgp-sets');
 
       const res = await request(app).get('/api/tcgdex/tcgp-sets/');
       expect(res.status).toBe(200);
@@ -111,12 +111,12 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     });
 
     it('handles select/orderBy combinations and invalid regex', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/tcgp-sets', cache, 'tcgpSets', [
         { name: 'Delta Set', code: 'D1', releaseDate: '2018-03-03', cardCount: 5 },
         { name: 'Epsilon Set', code: 'E1', releaseDate: '2022-02-02', cardCount: 15 },
         { name: 'Gamma Collection', code: 'G1', releaseDate: '2019-07-07', cardCount: 10 },
-      ], '../routes/tcgp-sets');
+      ], '../src/routes/tcgp-sets');
 
       const res = await request(app).get('/api/tcgdex/tcgp-sets').query({ select: 'name,nonexistent' });
       expect(res.status).toBe(200);
@@ -132,10 +132,10 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     });
 
     it('returns 500 when tcgp cache is undefined', async () => {
-      const cache = require('../cache');
+      const cache = require('../src/lib/cache');
       const app = makeAppWithRoute('/api/tcgdex/tcgp-sets', cache, 'tcgpSets', [
         { name: 'Delta Set', code: 'D1' }
-      ], '../routes/tcgp-sets');
+      ], '../src/routes/tcgp-sets');
 
       cache.tcgpSets = undefined;
       const res = await request(app).get('/api/tcgdex/tcgp-sets/');
@@ -145,10 +145,10 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     it('logs and returns 500 when handler throws for tcgp', async () => {
       jest.resetModules();
       const mockLogger = { error: jest.fn(), log: jest.fn() };
-      jest.doMock('../logger', () => mockLogger);
+      jest.doMock('../src/lib/logger', () => mockLogger);
 
-      const cache = require('../cache');
-      const app = makeAppWithRoute('/api/tcgdex/tcgp-sets', cache, 'tcgpSets', [ { name: 'X' } ], '../routes/tcgp-sets');
+      const cache = require('../src/lib/cache');
+      const app = makeAppWithRoute('/api/tcgdex/tcgp-sets', cache, 'tcgpSets', [ { name: 'X' } ], '../src/routes/tcgp-sets');
 
       Object.defineProperty(cache, 'tcgpSets', { get: () => { throw new Error('boom tcgp'); } });
 
@@ -162,10 +162,10 @@ describe('PTCG and TCGP sets routes (consolidated)', () => {
     it('logs and returns 500 when handler throws for ptcg', async () => {
       jest.resetModules();
       const mockLogger = { error: jest.fn(), log: jest.fn() };
-      jest.doMock('../logger', () => mockLogger);
+      jest.doMock('../src/lib/logger', () => mockLogger);
 
-      const cache = require('../cache');
-      const app = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [ { name: 'Y' } ], '../routes/ptcg-sets');
+      const cache = require('../src/lib/cache');
+      const app = makeAppWithRoute('/api/tcgdex/ptcg-sets', cache, 'ptcgSets', [ { name: 'Y' } ], '../src/routes/ptcg-sets');
 
       Object.defineProperty(cache, 'ptcgSets', { get: () => { throw new Error('boom ptcg'); } });
